Add order review

+++ Sources/App/Migration/CreateAddress.swift
@@ -5,7 +5,7 @@ struct CreateAddress: AsyncMigration {
         try await database.schema(Address.schema).id()
             .field("street", .string, .required)
             .field("postal_code_suffix", .string, .required)
-            .field("house_number", .int, .required)
+            .field("house_number", .int16, .required)
             .field("latitude", .double, .required)
             .field("longitude", .double, .required)
             .field(
