+++ Sources/App/Migration/CreateDiscount.swift
@@ -4,7 +4,7 @@ struct CreateDiscount: AsyncMigration {
     func prepare(on database: Database) async throws {
         try await database.schema(Discount.schema).id()
             .field("name", .string, .required)
-            .field("percentage", .int, .required)
+            .field("percentage", .int16, .required)
             .field("online_date", .datetime, .required)
             .field("offline_date", .datetime, .required)
             .field(
